forward pass working

+++ projects/StreamPETR/stream_petr/core/bbox/match_costs/match_cost.py
@@ -22,6 +22,6 @@ class BBox3DL1Cost(object):
         Returns:
             torch.Tensor: bbox_cost value with weight
         """
-        bbox_cost = torch.cdist(bbox_pred, gt_bboxes, p=1)
+        bbox_cost = torch.cdist(bbox_pred.float(), gt_bboxes.float(), p=1)
         return bbox_cost * self.weight
